# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Quran Hifdh Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
          description: Filter by active status
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentRequest"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Student:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        currentSurah:
          type: integer
        currentAyah:
          type: integer
        startDate:
          type: string
          format: date
        notes:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - currentSurah
        - currentAyah
        - startDate
        - active
        - createdAt
    CreateStudentRequest:
      type: object
      properties:
        name:
          type: string
        currentSurah:
          type: integer
        currentAyah:
          type: integer
        startDate:
          type: string
          format: date
        notes:
          type: string
          nullable: true
      required:
        - name
        - currentSurah
        - currentAyah
        - startDate
    UpdateStudentRequest:
      type: object
      properties:
        name:
          type: string
        currentSurah:
          type: integer
        currentAyah:
          type: integer
        notes:
          type: string
          nullable: true
        active:
          type: boolean
...