# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Hifdh Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  currentSurah: number;
  currentAyah: number;
  startDate: Date;
  notes?: string | null;
  active: boolean;
  createdAt: Date;
}